pyproject.toml
[project]
name = "mosaic"
version = "0.1.0"
description = "Public code release for MOSAIC."
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "absl-py>=2.4.0",
    "immutabledict>=4.3.1",
    "langdetect>=1.0.9",
    "matplotlib",
    "nest-asyncio>=1.6.0",
    "nltk>=3.9.2",
    "openai>=2.17.0",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "pydantic-ai>=1.61.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
]

[project.optional-dependencies]
# Training and managed inference:
#   uv sync --extra finetune
# Core evaluation / orchestration only:
#   uv sync
finetune = [
    "accelerate>=1.12.0",
    "datasets>=4.5.0",
    "peft>=0.18.1",
    "safetensors>=0.7.0",
    "sentencepiece>=0.2.1",
    "torch>=2.9.0,<2.10",
    "transformers>=4.56.0,<5",
    "trl>=0.28.0",
    "vllm>=0.15.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
